Write Accrue Interest Trigger Tests
Wrote Trigger Test for Accrue Interest. (#8)
Found a Nasty Bug with Data Format of DECIMAL in MySQL and fixed it.

+++ sql/utilities/Add_Money_To_Market_Account.sql
@@ -1,4 +1,4 @@
-CREATE PROCEDURE IF NOT EXISTS Add_Money_To_Market_Account (IN in_aid INT, IN amt DECIMAL)
+CREATE PROCEDURE IF NOT EXISTS Add_Money_To_Market_Account (IN in_aid INT, IN amt DECIMAL(18,3))
     UPDATE Market_Accounts
     SET balance = balance + amt
     WHERE aid = in_aid;
